Renaming module name
